Templating the ressource class to be able able to define int valued ressources and double valued ressources

+++ pathwyse/core/resources/defaultcost.cpp
@@ -11,19 +11,19 @@ DefaultCost::DefaultCost(double init_value) {
     this->init_value = init_value;
 }
 
-int DefaultCost::extend(int current_value, int i, int j, bool direction) {
+double DefaultCost::extend(double current_value, int i, int j, bool direction) {
     int dest_node = direction ? j : i;
     return current_value + data->getArcCost(i, j) + data->getNodeCost(dest_node);
 }
 
-int DefaultCost::join(int current_value_forward, int current_value_backward, int i, int j){
+double DefaultCost::join(double current_value_forward, double current_value_backward, int i, int j){
     return current_value_forward + data->getArcCost(i, j) + current_value_backward;
 }
 
-int DefaultCost::join(int current_value_forward, int current_value_backward, int node){
+double DefaultCost::join(double current_value_forward, double current_value_backward, int node){
     return current_value_forward + current_value_backward - data->getNodeCost(node);
 }
 
-bool DefaultCost::isFeasible(int current_value, int current_node, double bounding, bool direction) {
+bool DefaultCost::isFeasible(double current_value, int current_node, double bounding, bool direction) {
     return current_value <= upper_bound;
 }
